+++ tests/unit/Model/Command/UserMergeTest.php
@@ -2,6 +2,7 @@
 
 namespace Lmc\Matej\Model\Command;
 
+use Assert\InvalidArgumentException;
 use Lmc\Matej\UnitTestCase;
 
 class UserMergeTest extends UnitTestCase
@@ -17,6 +18,14 @@ class UserMergeTest extends UnitTestCase
 
         $command = UserMerge::mergeFromSourceToTargetUser($sourceUserId, $targetUserId);
         $this->assertUserMergeCommand($command, $sourceUserId, $targetUserId);
+    }
+
+    public function shouldThrowExceptionWhenMergingSameUsers(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('You have to provide different source and target user id in UserMerge ("test-user" set for both)');
+
+        UserMerge::mergeInto('test-user', 'test-user');
     }
 
     /**